server/requirements.txt
Flask
mysql-connector-python
python-dotenv
mysql.connector
flask_cors
pandas
moralis